+++ .gitignore
@@ -14,8 +14,5 @@ data
 .env
 .env.server
 
-# elf
-**elf
-
 # txt
 **.txt
