+++ client-programs/aggregation/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "aggregation"
+version = "0.1.0"
+license.workspace = true
+edition.workspace = true
+authors.workspace = true
+homepage.workspace = true
+repository.workspace = true
+
+[dependencies]
+sha2 = { workspace = true }
+sp1-zkvm = { workspace = true }
+sp1-lib = { workspace = true }
+client-utils = { workspace = true }
+alloy-consensus = { workspace = true }
+alloy-primitives = { workspace = true }
+serde_cbor.workspace = true
