+++ client-programs/fault-proof/README.md
@@ -1,0 +1,3 @@
+# `fault_proof`
+
+This binary contains the client program for executing the Optimism rollup state transition across a single block, which can be used in an on chain dispute game. Depending on the compilation pipeline, it will compile to be run on `kona-host` or on `zkvm-host` (in SP1).
