+++ crates/host-utils/src/fetcher.rs
@@ -331,10 +331,10 @@ impl SP1KonaDataFetcher {
         // The native programs are built with profile release-client-lto in build.rs
         let exec_directory = match multi_block {
             ProgramType::Single => {
-                format!("{}/target/release-client-lto/zkvm-client", workspace_root)
+                format!("{}/target/release-client-lto/fault_proof", workspace_root)
             }
             ProgramType::Multi => format!(
-                "{}/target/release-client-lto/validity-client",
+                "{}/target/release-client-lto/range",
                 workspace_root
             ),
         };
