+++ prover-server/bin/server.rs
@@ -18,8 +18,8 @@ use sp1_sdk::{
 };
 use std::{env, fs};
 
-pub const MULTI_BLOCK_ELF: &[u8] = include_bytes!("../../elf/validity-client-elf");
-pub const AGG_ELF: &[u8] = include_bytes!("../../elf/aggregation-client-elf");
+pub const MULTI_BLOCK_ELF: &[u8] = include_bytes!("../../elf/range-elf");
+pub const AGG_ELF: &[u8] = include_bytes!("../../elf/aggregation-elf");
 
 #[derive(Deserialize, Serialize, Debug)]
 struct SpanProofRequest {
