+++ zkvm-host/bin/agg.rs
@@ -10,8 +10,8 @@ use host_utils::{
 };
 use sp1_sdk::{utils, HashableKey, ProverClient, SP1Proof, SP1ProofWithPublicValues};
 
-pub const AGG_ELF: &[u8] = include_bytes!("../../elf/aggregation-client-elf");
-pub const MULTI_BLOCK_ELF: &[u8] = include_bytes!("../../elf/validity-client-elf");
+pub const AGG_ELF: &[u8] = include_bytes!("../../elf/aggregation-elf");
+pub const MULTI_BLOCK_ELF: &[u8] = include_bytes!("../../elf/range-elf");
 
 #[derive(Parser, Debug)]
 #[command(author, version, about, long_about = None)]
